# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MetalPrice {
  /** Metal ticker, e.g. XAU, XAG */
  symbol: string;
  name: string;
  price: number;
  currency: string;
}

export interface ForexRate {
  /** Currency pair, e.g. EUR/USD */
  pair: string;
  rate: number;
}

export interface PriceSnapshot {
  updatedAt: string;
  metals: MetalPrice[];
  forex: ForexRate[];
  /** True if the last upstream price fetch failed and this is a cached snapshot */
  stale: boolean;
}

export type AlertDirection = typeof AlertDirection[keyof typeof AlertDirection];


export const AlertDirection = {
  above: 'above',
  below: 'below',
} as const;

export type AlertStatus = typeof AlertStatus[keyof typeof AlertStatus];


export const AlertStatus = {
  active: 'active',
  triggered: 'triggered',
  acknowledged: 'acknowledged',
  disabled: 'disabled',
} as const;

export interface Alert {
  id: number;
  /** XAU, XAG, EUR/USD, GBP/USD, USD/JPY */
  assetSymbol: string;
  /** Human readable label, e.g. Gold (XAU) */
  assetLabel: string;
  targetPrice: number;
  direction: AlertDirection;
  status: AlertStatus;
  /** @nullable */
  note: string | null;
  createdAt: string;
  /** @nullable */
  triggeredAt: string | null;
  /** @nullable */
  acknowledgedAt: string | null;
}

export type AlertInputDirection = typeof AlertInputDirection[keyof typeof AlertInputDirection];


export const AlertInputDirection = {
  above: 'above',
  below: 'below',
} as const;

export interface AlertInput {
  assetSymbol: string;
  assetLabel: string;
  targetPrice: number;
  direction: AlertInputDirection;
  /** @nullable */
  note?: string | null;
}

export type AlertUpdateDirection = typeof AlertUpdateDirection[keyof typeof AlertUpdateDirection];


export const AlertUpdateDirection = {
  above: 'above',
  below: 'below',
} as const;

export type AlertUpdateStatus = typeof AlertUpdateStatus[keyof typeof AlertUpdateStatus];


export const AlertUpdateStatus = {
  active: 'active',
  disabled: 'disabled',
} as const;

export interface AlertUpdate {
  targetPrice?: number;
  direction?: AlertUpdateDirection;
  /** @nullable */
  note?: string | null;
  status?: AlertUpdateStatus;
}

export interface AccountStatus {
  isPremium: boolean;
  trialStartedAt: string;
  trialEndsAt: string;
  /** Days left in the free trial, 0 if expired or premium */
  daysRemaining: number;
  canCreateAlerts: boolean;
}

export interface SubscriptionRequiredError {
  error: string;
  contactUrl: string;
}