# lib/api-zod/src/generated/types/accountStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccountStatus {
  isPremium: boolean;
  trialStartedAt: Date;
  trialEndsAt: Date;
  /** Days left in the free trial, 0 if expired or premium */
  daysRemaining: number;
  canCreateAlerts: boolean;
}