# lib/api-zod/src/generated/types/subscriptionRequiredError.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriptionRequiredError {
  error: string;
  contactUrl: string;
}